requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==3.0.1
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
colorama==0.4.6
discord.py==1.7.3
distlib==0.3.6
Django==4.2.2
django-cors-headers==4.0.0
et-xmlfile==1.1.0
filelock==3.12.0
frozenlist==1.3.3
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
load-dotenv==0.1.0
multidict==6.0.4
mysqlclient==2.1.1
numpy==1.24.3
openai==0.27.7
openpyxl==3.1.2
pandas==2.0.2
platformdirs==3.5.1
pycparser==2.21
PyNaCl==1.5.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytube==15.0.0
pytz==2023.3
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
virtualenv==20.23.0
yarl==1.9.2
youtube-dl==2021.12.17